# Conflicts:
#	src/main/java/ru/technoserv/dao/impl/HibernateEmployeeDao.java

+++ src/main/java/ru/technoserv/services/impl/EmployeeServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.technoserv.dao.*;
+import ru.technoserv.domain.Department;
 import ru.technoserv.domain.Employee;
 import ru.technoserv.domain.EmployeeHistory;
 import ru.technoserv.exceptions.EmployeeTheHeadOfDepartment;
@@ -33,7 +34,6 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Override
     public Employee createEmployee(Employee employee) {
-     //   log.info("Посылаем запрос dao на создание сотрудника: "+employee);
         logger.info("Создаем сотрудника");
 
         EmployeeHistory eh = new EmployeeHistory(employee);
@@ -54,6 +54,13 @@ public class EmployeeServiceImpl implements EmployeeService {
     @Override
     public void removeEmployee(int id) {
         logger.info("Удаляем сотрудника");
+        Department empDept = employeeDao.read(id).getDepartment();
+        int empDeptHeadID = empDept.getDeptHeadId();
+
+        if (empDeptHeadID == id) {
+            logger.info("Попытка удалить начальника отдела. Действие невозможно");
+            throw new EmployeeTheHeadOfDepartment(id);
+        }
         employeeDao.delete(id);
     }
 
@@ -80,7 +87,6 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     public List<Employee> getEmployees(int depID){
         logger.info("Получаем сотрудников по отделу");
-      //  log.info("Посылаем запрос dao на получение сотрудников отдела с ИД "+depID);
         List<EmployeeHistory> allEmpsHistory = employeeDao.getAllFromDept(depID);
         for(EmployeeHistory eh : allEmpsHistory) {
             eh.setDepartment(departmentDao.readById(depID));
